+++ Sources/Sentry/SentryNetworkTracker.m
@@ -8,6 +8,8 @@
 #import "SentryScope+Private.h"
 #import "SentrySpan.h"
 #import "SentryTraceHeader.h"
+#import "SentryTraceState.h"
+#import "SentryTracer.h"
 #import <objc/runtime.h>
 
 static NSString *const SENTRY_NETWORK_REQUEST_TRACKER_SPAN = @"SENTRY_NETWORK_REQUEST_TRACKER_SPAN";
@@ -252,13 +254,15 @@ SentryNetworkTracker ()
         return headers;
     }
 
-    if (headers == nil) {
-        return @{ SENTRY_TRACE_HEADER : [span toTraceHeader].value };
-    } else {
-        NSMutableDictionary *newHeaders = [[NSMutableDictionary alloc] initWithDictionary:headers];
-        newHeaders[SENTRY_TRACE_HEADER] = [span toTraceHeader].value;
-        return [[NSDictionary alloc] initWithDictionary:newHeaders];
+    NSMutableDictionary *result = [[NSMutableDictionary alloc] initWithDictionary:headers];
+    result[SENTRY_TRACE_HEADER] = [span toTraceHeader].value;
+
+    SentryTracer *tracer = [SentryTracer getTracer:span];
+    if (tracer != nil) {
+        result[SENTRY_TRACESTATE_HEADER] = [tracer.traceState toHTTPHeader];
     }
+
+    return [[NSDictionary alloc] initWithDictionary:result];
 }
 
 @end