+++ Tests/SentryTests/SentryOptionsTest.m
@@ -407,6 +407,24 @@
     XCTAssertEqual(NO, options.stitchAsyncCode);
 }
 
+- (void)testEnableTraceSampling
+{
+    SentryOptions *options = [self getValidOptions:@{}];
+    XCTAssertFalse(options.experimentalEnableTraceSampling);
+}
+
+- (void)testEnableTraceSamplingEnabled
+{
+    SentryOptions *options = [self getValidOptions:@{ @"experimentalEnableTraceSampling" : @YES }];
+    XCTAssertTrue(options.experimentalEnableTraceSampling);
+}
+
+- (void)testEnableTraceSamplingDisabled
+{
+    SentryOptions *options = [self getValidOptions:@{ @"experimentalEnableTraceSampling" : @NO }];
+    XCTAssertFalse(options.experimentalEnableTraceSampling);
+}
+
 - (void)testStitchAsyncCodeEnabled
 {
     SentryOptions *options = [self getValidOptions:@{ @"stitchAsyncCode" : @YES }];
@@ -439,6 +457,7 @@
     XCTAssertEqual(NO, options.stitchAsyncCode);
     XCTAssertEqual(20 * 1024 * 1024, options.maxAttachmentSize);
     XCTAssertTrue(options.enableAutoPerformanceTracking);
+    XCTAssertFalse(options.experimentalEnableTraceSampling);
 }
 
 - (void)testSetValidDsn